# Serveur/public/swagger.json
{
	"components": {
...
		"schemas": {
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"IAnimal": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"name": {
						"type": "string"
					},
					"birthdate": {
						"type": "string",
						"format": "date-time"
					},
					"speciesID": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"userID": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"trackerID": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"birthdate",
					"speciesID",
					"userID",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPoint": {
				"properties": {
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ZONE_TYPE": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
			"IZone": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"color": {
						"type": "string"
					},
					"points": {
						"items": {
							"$ref": "#/components/schemas/IPoint"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"userId": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"type": {
						"$ref": "#/components/schemas/ZONE_TYPE"
					},
					"speciesId": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"animalsId": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"agricoleStatId": {
						"$ref": "#/components/schemas/ObjectId"
					}
				},
				"required": [
					"_id",
					"color",
					"points",
					"name",
					"userId",
					"type",
					"speciesId",
					"animalsId",
					"agricoleStatId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAgricoleStat": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"userId": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"humidity": {
						"type": "number",
						"format": "double"
					},
					"sun": {
						"type": "number",
						"format": "double"
					},
					"airPurity": {
						"type": "number",
						"format": "double"
					},
					"lastRecolt": {
						"type": "string",
						"format": "date-time"
					},
					"lastProduit": {
						"type": "string"
					},
					"typeAgriculture": {
						"type": "string"
					},
					"vegetable": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"userId",
					"humidity",
					"sun",
					"airPurity",
					"lastRecolt",
					"lastProduit",
					"typeAgriculture",
					"vegetable"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IAgricoleStat_": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"userId": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"humidity": {
						"type": "number",
						"format": "double"
					},
					"sun": {
						"type": "number",
						"format": "double"
					},
					"airPurity": {
						"type": "number",
						"format": "double"
					},
					"lastRecolt": {
						"type": "string",
						"format": "date-time"
					},
					"lastProduit": {
						"type": "string"
					},
					"typeAgriculture": {
						"type": "string"
					},
					"vegetable": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "animalia-backend",
		"version": "1.0.0",
		"description": "Le backend du serveur e Animalia ",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Lucas Oulieu & Leïna Sadi"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/animal/zone/{zoneId}": {
			"get": {
				"operationId": "GetByZone",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IAnimal"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Animals"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "zoneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/zone/{_id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IZone"
								}
							}
						}
					}
				},
				"tags": [
					"Zone"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/agricole/user/{userId}": {
			"get": {
				"operationId": "Gets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IAgricoleStat"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Agricole Stat"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/agricole/zone/{zoneId}": {
			"get": {
				"operationId": "GetByZone",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAgricoleStat"
								}
							}
						}
					}
				},
				"tags": [
					"Agricole Stat"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "zoneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/agricole/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAgricoleStat"
								}
							}
						}
					}
				},
				"tags": [
					"Agricole Stat"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Partial_IAgricoleStat_"
									},
									{
										"properties": {
											"_id": {
												"type": "string"
											}
										},
										"required": [
											"_id"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}